Fix references

+++ src/Orckestra.Composer.Cart/Services/CartService.cs
@@ -43,7 +43,6 @@ namespace Orckestra.Composer.Cart.Services
         protected IImageService ImageService { get; private set; }
         protected ICategoryRepository CategoryRepository { get; private set; }
         protected ILocalizationProvider LocalizationProvider { get; private set; }
-        protected IComposerContext ComposerContext { get; private set; }
 
         /// <summary>
         /// CartService constructor
@@ -59,19 +58,18 @@ namespace Orckestra.Composer.Cart.Services
         /// <param name="regionCodeProvider">The <see cref="IRegionCodeProvider"/></param>
         /// <param name="imageService">The <see cref="IImageService"/></param>
         public CartService(
-            ICartRepository cartRepository,
-            IDamProvider damProvider,
+            ICartRepository cartRepository, 
+            IDamProvider damProvider, 
             ICartViewModelFactory cartViewModelFactory,
             ICouponViewService couponViewService,
-            ILookupService lookupService,
+            ILookupService lookupService, 
             ILineItemService lineItemService,
             IFixCartService fixCartService,
             ICountryService countryService,
             IRegionCodeProvider regionCodeProvider,
             IImageService imageService,
             ICategoryRepository categoryRepository,
-            ILocalizationProvider localizationProvider,
-            IComposerContext composerContext)
+            ILocalizationProvider localizationProvider)
         {
             CartRepository = cartRepository ?? throw new ArgumentNullException(nameof(cartRepository));
             CartViewModelFactory = cartViewModelFactory ?? throw new ArgumentNullException(nameof(cartViewModelFactory));
@@ -83,7 +81,6 @@ namespace Orckestra.Composer.Cart.Services
             RegionCodeProvider = regionCodeProvider ?? throw new ArgumentNullException(nameof(regionCodeProvider));
             ImageService = imageService ?? throw new ArgumentNullException(nameof(imageService));
             LocalizationProvider = localizationProvider ?? throw new ArgumentNullException(nameof(localizationProvider));
-            ComposerContext = composerContext ?? throw new ArgumentNullException(nameof(composerContext));
             CategoryRepository = categoryRepository ?? throw new ArgumentNullException(nameof(categoryRepository));
         }
 
@@ -206,7 +203,7 @@ namespace Orckestra.Composer.Cart.Services
             };
 
             var viewModel = await CreateCartViewModelAsync(vmParam).ConfigureAwait(false);
-
+            
             return viewModel;
         }
 
@@ -225,10 +222,10 @@ namespace Orckestra.Composer.Cart.Services
             if (string.IsNullOrWhiteSpace(param.BaseUrl)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.BaseUrl)), nameof(param)); }
             if (string.IsNullOrWhiteSpace(param.BillingCurrency)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.BillingCurrency)), nameof(param)); }
             if (string.IsNullOrWhiteSpace(param.CartName)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.CartName)), nameof(param)); }
+            if (string.IsNullOrWhiteSpace(param.CartType)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.CartType)), nameof(param)); }
             if (string.IsNullOrWhiteSpace(param.Scope)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.Scope)), nameof(param)); }
-            if (string.IsNullOrWhiteSpace(param.Status)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.Status)), nameof(param)); }
+            if (string.IsNullOrWhiteSpace(param.Status)){ throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.Status)), nameof(param)); }
 
-            param.CartType = GetCurrentCartType();
             var updatedCart = await CartRepository.UpdateCartAsync(param).ConfigureAwait(false);
 
             await CartRepository.RemoveCouponsAsync(new RemoveCouponsParam
@@ -248,7 +245,7 @@ namespace Orckestra.Composer.Cart.Services
             };
 
             var viewModel = await CreateCartViewModelAsync(vmParam).ConfigureAwait(false);
-
+            
             return viewModel;
         }
 
@@ -265,12 +262,7 @@ namespace Orckestra.Composer.Cart.Services
             if (string.IsNullOrWhiteSpace(param.CartName)) { throw new ArgumentException(GetMessageOfNullWhiteSpace(nameof(param.CartName)), nameof(param)); }
             if (param.CustomerId == Guid.Empty) { throw new ArgumentException(GetMessageOfEmpty(nameof(param.CustomerId)), nameof(param)); }
 
-            var cart = await CartRepository.GetCartAsync(param).ConfigureAwait(false);
-
-            if (!ComposerContext.IsEditingOrder) // Don't FixCart while editing order
-            {
-                await FixCartService.FixCartAsync(new FixCartParam { Cart = cart });
-            }
+            var cart = await CartRepository.GetCartAsync(param).ConfigureAwait(false);            
 
             await CartRepository.RemoveCouponsAsync(new RemoveCouponsParam
             {
@@ -317,7 +309,7 @@ namespace Orckestra.Composer.Cart.Services
 
             var vm = CartViewModelFactory.CreateCartViewModel(param);
 
-            if (CartConfiguration.GroupCartItemsByPrimaryCategory)
+            if(CartConfiguration.GroupCartItemsByPrimaryCategory)
             {
                 vm.GroupedLineItemDetailViewModels = await GetGroupedLineItems(vm, param).ConfigureAwait(false);
             }
@@ -464,7 +456,7 @@ namespace Orckestra.Composer.Cart.Services
             }).ConfigureAwait(false);
 
             country.Validate(param.PostalCode);
-
+            
             if (shipment.Address == null)
             {
                 shipment.Address = new Address { PropertyBag = new PropertyBag() };
@@ -494,7 +486,7 @@ namespace Orckestra.Composer.Cart.Services
 
             }).ConfigureAwait(false);
 
-            var payment = cart.Payments?.Find(x => !x.IsVoided())
+            var payment = cart.Payments?.Find(x => !x.IsVoided()) 
             ?? throw new InvalidOperationException("There is no valid payment from which we can get or set the billing address");
 
             if (payment.BillingAddress == null)
@@ -503,7 +495,7 @@ namespace Orckestra.Composer.Cart.Services
             }
 
             await MapBillingAddressPostalCodeToPaymentAsync(param, payment);
-
+            
             return await UpdateCartAsync(new UpdateCartViewModelParam
             {
                 BaseUrl = param.BaseUrl,
@@ -559,7 +551,7 @@ namespace Orckestra.Composer.Cart.Services
 
             country.Validate(param.PostalCode);
 
-            if (payment.BillingAddress.PropertyBag != null)
+            if(payment.BillingAddress.PropertyBag != null)
                 payment.BillingAddress.PropertyBag[AddressBookIdPropertyBagKey] = Guid.Empty; // because the updated address will not correspond to any registered address
             payment.BillingAddress.PostalCode = param.PostalCode;
             payment.BillingAddress.CountryCode = country.IsoCode;
@@ -568,16 +560,9 @@ namespace Orckestra.Composer.Cart.Services
 
         protected virtual string GetRegionCodeBasedOnPostalCode(string postalCode, string countryCode)
         {
-            var region = RegionCodeProvider.GetRegion(postalCode, countryCode) ??
+            var region = RegionCodeProvider.GetRegion(postalCode, countryCode) ?? 
                 throw new ArgumentException("Cannot resolve a Region based on this Postal Code", "postalCode");
             return region;
-        }
-
-        private string GetCurrentCartType()
-        {
-            return !ComposerContext.IsEditingOrder
-                        ? CartConfiguration.DefaultCartType
-                        : CartConfiguration.EditingCartType;
         }
     }
 }
